services/ollama/requirements.txt
flask
flask_cors
requests
openai
ollama  # Ollama's Python package (if available)
pinecone-client
